+++ tests/test_manager.py
@@ -1,0 +1,10 @@
+import copy
+import numpy as np
+from unittest import TestCase
+from NoveltyProducer.Manager import Manager
+
+class TestBaseUnit(TestCase):
+
+    def test_type_output(self):
+        """Test types of output"""
+        pass